Refactor: modularised ServerToClient socket handler & queue
